extract common models/functions
