+++ .github/workflows/_build-image.yml
@@ -1,0 +1,104 @@
+name: Build and Push Image (Reusable)
+
+on:
+  workflow_call:
+    inputs:
+      image_name:
+        required: true
+        type: string
+        description: "ECR image name (e.g., p3v1o3c6/registry)"
+      dockerfile:
+        required: true
+        type: string
+        description: "Path to Dockerfile (e.g., docker/Dockerfile.registry)"
+      timeout_minutes:
+        required: false
+        type: number
+        default: 45
+        description: "Job timeout in minutes"
+      role_session_name:
+        required: true
+        type: string
+        description: "AWS role session name (e.g., registry-build)"
+      summary_title:
+        required: true
+        type: string
+        description: "Title for the GitHub Step Summary (e.g., Registry Image Published)"
+      extra_build_args:
+        required: false
+        type: string
+        default: ""
+        description: "Additional Docker build args (newline-separated)"
+    secrets:
+      ECR_ROLE:
+        required: true
+
+env:
+  REGISTRY: public.ecr.aws
+
+jobs:
+  build-and-push:
+    if: github.repository == 'agentic-community/mcp-gateway-registry'
+    name: Build and Push
+    runs-on: ubuntu-latest
+    timeout-minutes: ${{ inputs.timeout_minutes }}
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@8d2750c68a42422c14e847fe6c8ac0403b4cbd6f # v3.12.0
+
+      - name: Configure Role to Acquire Credentials
+        uses: aws-actions/configure-aws-credentials@8df5847569e6427dd6c4fb1cf565c83acfa8afa7 # v6.0.0
+        with:
+          aws-region: us-east-1
+          role-session-name: ${{ inputs.role_session_name }}
+          role-to-assume: ${{ secrets.ECR_ROLE }}
+
+      - name: Login to Amazon ECR Public
+        id: login-ecr-public
+        uses: aws-actions/amazon-ecr-login@062b18b96a7aff071d4dc91bc00c4c1a7945b076 # v2.0.1
+        with:
+          registry-type: public
+
+      - name: Extract metadata
+        id: meta
+        uses: docker/metadata-action@c299e40c65443455700f0fdfc63efafe5b349051 # v5.10.0
+        with:
+          images: ${{ env.REGISTRY }}/${{ inputs.image_name }}
+          tags: |
+            type=raw,value=latest,enable={{is_default_branch}}
+            type=sha,prefix=,format=long
+
+      - name: Build and push
+        id: push
+        uses: docker/build-push-action@263435318d21b8e681c14492fe198d362a7d2c83 # v6.18.0
+        with:
+          context: .
+          file: ${{ inputs.dockerfile }}
+          push: true
+          tags: ${{ steps.meta.outputs.tags }}
+          labels: ${{ steps.meta.outputs.labels }}
+          build-args: |
+            BUILD_VERSION=${{ github.sha }}
+            ${{ inputs.extra_build_args }}
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
+
+      - name: Generate attestation
+        uses: actions/attest-build-provenance@96278af6caaf10aea03fd8d33a09a777ca52d62f # v3.2.0
+        with:
+          subject-name: ${{ env.REGISTRY }}/${{ inputs.image_name }}
+          subject-digest: ${{ steps.push.outputs.digest }}
+          push-to-registry: true
+
+      - name: Image Summary
+        run: |
+          echo "## ${{ inputs.summary_title }}" >> $GITHUB_STEP_SUMMARY
+          echo "" >> $GITHUB_STEP_SUMMARY
+          echo "**Tags:**" >> $GITHUB_STEP_SUMMARY
+          echo '```' >> $GITHUB_STEP_SUMMARY
+          echo "${{ steps.meta.outputs.tags }}" >> $GITHUB_STEP_SUMMARY
+          echo '```' >> $GITHUB_STEP_SUMMARY
