+++ .github/workflows/build-mcpgw.yml
@@ -6,6 +6,7 @@ on:
     paths:
       - 'servers/mcpgw/**'
       - 'registry/**'
+      - 'common/**'
       - 'docker/Dockerfile.mcp-server'
       - '.github/workflows/build-mcpgw.yml'
   workflow_dispatch:
@@ -20,73 +21,15 @@ concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
   cancel-in-progress: true
 
-env:
-  REGISTRY: public.ecr.aws
-  IMAGE_NAME: p3v1o3c6/mcpgw
-
 jobs:
   build-and-push:
-    if: github.repository == 'agentic-community/mcp-gateway-registry'
-    name: Build and Push
-    runs-on: ubuntu-latest
-    timeout-minutes: 45
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@8d2750c68a42422c14e847fe6c8ac0403b4cbd6f # v3.12.0
-
-      - name: Configure Role to Acquire Credentials
-        uses: aws-actions/configure-aws-credentials@8df5847569e6427dd6c4fb1cf565c83acfa8afa7 # v6.0.0
-        with:
-          aws-region: us-east-1
-          role-session-name: mcpgw-build
-          role-to-assume: ${{ secrets.ECR_ROLE }}
-
-      - name: Login to Amazon ECR Public
-        id: login-ecr-public
-        uses: aws-actions/amazon-ecr-login@062b18b96a7aff071d4dc91bc00c4c1a7945b076 # v2.0.1
-        with:
-          registry-type: public
-
-      - name: Extract metadata
-        id: meta
-        uses: docker/metadata-action@c299e40c65443455700f0fdfc63efafe5b349051 # v5.10.0
-        with:
-          images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
-          tags: |
-            type=raw,value=latest,enable={{is_default_branch}}
-            type=sha,prefix=,format=long
-
-      - name: Build and push
-        id: push
-        uses: docker/build-push-action@263435318d21b8e681c14492fe198d362a7d2c83 # v6.18.0
-        with:
-          context: .
-          file: docker/Dockerfile.mcp-server
-          push: true
-          tags: ${{ steps.meta.outputs.tags }}
-          labels: ${{ steps.meta.outputs.labels }}
-          build-args: |
-            BUILD_VERSION=${{ github.sha }}
-            SERVER_DIR=servers/mcpgw
-          cache-from: type=gha
-          cache-to: type=gha,mode=max
-
-      - name: Generate attestation
-        uses: actions/attest-build-provenance@96278af6caaf10aea03fd8d33a09a777ca52d62f # v3.2.0
-        with:
-          subject-name: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
-          subject-digest: ${{ steps.push.outputs.digest }}
-          push-to-registry: true
-
-      - name: Image Summary
-        run: |
-          echo "## MCPGW Image Published" >> $GITHUB_STEP_SUMMARY
-          echo "" >> $GITHUB_STEP_SUMMARY
-          echo "**Tags:**" >> $GITHUB_STEP_SUMMARY
-          echo '```' >> $GITHUB_STEP_SUMMARY
-          echo "${{ steps.meta.outputs.tags }}" >> $GITHUB_STEP_SUMMARY
-          echo '```' >> $GITHUB_STEP_SUMMARY
+    uses: ./.github/workflows/_build-image.yml
+    with:
+      image_name: p3v1o3c6/mcpgw
+      dockerfile: docker/Dockerfile.mcp-server
+      timeout_minutes: 45
+      role_session_name: mcpgw-build
+      summary_title: MCPGW Image Published
+      extra_build_args: SERVER_DIR=servers/mcpgw
+    secrets:
+      ECR_ROLE: ${{ secrets.ECR_ROLE }}
