+++ common/__init__.py
@@ -1,0 +1,5 @@
+"""Common shared package for MCP Gateway Registry services.
+
+This package provides shared utilities, models, and configurations
+used across multiple services (registry, auth_server, mcpgw, agents, api).
+"""
