+++ common/auth/__init__.py
@@ -1,0 +1,1 @@
+"""Shared authentication utilities for MCP Gateway Registry services."""
