+++ common/models/__init__.py
@@ -1,0 +1,1 @@
+"""Shared Pydantic models used across MCP Gateway Registry services."""
