+++ common/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Shared utilities for MCP Gateway Registry services."""
